Update README.md
•	Number Guessing Game
This is a simple Python program that allows the user to play a number guessing game. The program generates a random number between 1 and 100, and the user has to guess the number. After each guess, the program provides feedback on whether the guess was too low or too high. The game continues until the user correctly guesses the number or decides to quit.


•	How to use it
To use the program, simply run the Python file. The program will prompt the user to enter their name and then start the game. The user can enter their guesses one at a time. If the user wants to quit the game, they can enter "quit" instead of a number. After each guess, the program will tell the user whether the guess was too high or too low. When the user correctly guesses the number, the program will display the number of guesses it took and whether the user set a new high score.


Requirements and Dependencies
This program requires Python 3.x to run. It also uses the following Python modules:

1.	random
2.	time
3.	winsound (only on Windows)


•	Code Explanation
play_game() function
The play_game() function is the main function that runs the number guessing game. It starts by welcoming the user and asking for their name. It then enters a loop that generates a new random number each time it runs. The user can guess the number by entering a guess one at a time. If the user enters "quit" instead of a number, the function will exit the loop and end the game. The function also handles invalid input and provides feedback to the user on whether their guess was too high or too low. When the user correctly guesses the number, the function calculates the time it took and displays the number of guesses and the time taken. If the user set a new high score, the function also displays a message congratulating the user on setting a new high score.

•	Winsound module
The winsound module is used to play sound effects when the user's guess is too high or too low, or when they set a new high score. This module is only available on Windows, so it may not work on other operating systems.

•	Variables
num: Stores the random number generated by the program.
num_guesses: Stores the number of guesses the user has made for the current round of the game.
high_score: Stores the lowest number of guesses it took the user to correctly guess the number in any round of the game. If the user beats the high score, this variable is updated to reflect the new high score.
start_time: Stores the time at which the user started guessing the number.
end_time: Stores the time at which the user correctly guessed the number.
time_taken: Stores the time taken by the user to correctly guess the number.